This is Rule 0:  a project with the goal of allowing PostgreSQL to fulfill Codd's 0th Rule for relational databases:  to "be able to manage data bases entirely through its relational capabilities".
As a bonus, it really should fulfill Rule 1, as well:  "All
information in a relational data base is represented explicitly
at the logical level and in exactly one way – by values in tables."

So to the extent that this project is successful, it would make SQL's
Data Definition Language redundant.  You should (for instance) be able
to create, alter, or delete a table via INSERT, UPDATE, and DROP
commands.  Also, Rule 0 should provide relational ways to interact with
users, permissions, and whatnot.

So far, a smallish hodge-podge of handy tools have been created and used
in an ad hoc fashion in the author's projects.  This initial commit
begins to try to make a coherent project out of them.